@@ -22,7 +22,8 @@ class Constant extends Token implements ProcessorInterface
      * Constant Processor walks through a Config structure and replaces all
      * PHP constants with their respective values
      *
-     * @param bool   $userOnly              True to process only user-defined constants, false to process all PHP constants
+     * @param bool   $userOnly              True to process only user-defined constants,
+     *                                      false to process all PHP constants
      * @param string $prefix                Optional prefix
      * @param string $suffix                Optional suffix
      * @return \Zend\Config\Processor\Constant
