@@ -9,6 +9,7 @@
 
 namespace ZendTest\Config\Reader;
 
+use Zend\Config\Exception;
 use Zend\Config\Reader\Ini;
 
 /**
@@ -35,7 +36,7 @@ class IniTest extends AbstractReaderTestCase
     public function testInvalidIniFile()
     {
         $this->reader = new Ini();
-        $this->setExpectedException('Zend\Config\Exception\RuntimeException');
+        $this->expectException(Exception\RuntimeException::class);
         $arrayIni = $this->reader->fromFile($this->getTestAssetPath('invalid'));
     }
 
@@ -60,7 +61,7 @@ ECS;
 test== "foo"
 
 ECS;
-        $this->setExpectedException('Zend\Config\Exception\RuntimeException');
+        $this->expectException(Exception\RuntimeException::class);
         $arrayIni = $this->reader->fromString($ini);
     }
 
